Quantile clean-up part 1: move bind data and sort tree to separate files
